move some files around #357
