+++ packages/cli/src/templates/plugin/app/components/pluginName/pluginName.tsx
@@ -1,10 +1,12 @@
 import { Component, h } from "@stencil/core";
 import pluginName from "../../../shared/pluginName";
 import store from "../../store/pluginName";
+
 @Component({
   tag: "cj-pluginName",
 })
 export class PluginName {
+
   render() {
     return (
       <corejam-box>
@@ -12,5 +14,6 @@ export class PluginName {
         {pluginName()}
       </corejam-box>
     );
+
   }
 }
