+++ packages/cli/src/templates/plugin/app/layout/layout.tsx
@@ -4,6 +4,7 @@ import { Component, h } from "@stencil/core";
   tag: "app-layout",
 })
 export class AppLayout {
+
   render() {
     return (
       <corejam-box>
@@ -11,4 +12,5 @@ export class AppLayout {
       </corejam-box>
     );
   }
+
 }
