+++ packages/cli/src/templates/plugin/app/routes/index.tsx
@@ -4,6 +4,7 @@ import { Component, h, Host } from "@stencil/core";
   tag: "route-index",
 })
 export class Hallo {
+
   render() {
     return (
       <Host>
@@ -16,4 +17,5 @@ export class Hallo {
       </Host>
     );
   }
+
 }
