+++ packages/cli/src/templates/plugin/shared/graphql/Mutations.ts
@@ -1,5 +1,4 @@
 
-
 export const pluginNameCreate = `
   mutation PluginNameCreate($pl: PluginNameCreateInput!) {
     pluginNameCreate(pluginNameCreateInput: $userCreateInput) {
