+++ packages/cli/src/templates/plugin/shared/graphql/Queries.ts
@@ -1,5 +1,4 @@
 
-
 export const allPluginNameGQL = `
   query {
     allPluginName {
