+++ packages/cli/src/templates/plugin/shared/types/pluginName.ts
